+++ pyfluids/config/__init__.py
@@ -1,7 +1,11 @@
 from .pyfluids_config import *
 from .pyfluids_config_builder import *
 from .unit_converter import *
+from .units_system import *
 
 __all__ = (
-    pyfluids_config.__all__ + pyfluids_config_builder.__all__ + unit_converter.__all__
+    pyfluids_config.__all__
+    + pyfluids_config_builder.__all__
+    + unit_converter.__all__
+    + units_system.__all__
 )
