+++ pyfluids/config/unit_converter.py
@@ -1,6 +1,6 @@
 from .pyfluids_config import PyFluidsConfig
 from .pyfluids_config_builder import PyFluidsConfigBuilder
-from ..enums import UnitsSystem
+from .units_system import UnitsSystem
 
 __all__ = ["UnitConverter"]
