+++ pyfluids/enums/__init__.py
@@ -1,6 +1,5 @@
 from .fluids_list import *
 from .mix import *
 from .phases import *
-from .units_system import *
 
-__all__ = fluids_list.__all__ + mix.__all__ + phases.__all__ + units_system.__all__
+__all__ = fluids_list.__all__ + mix.__all__ + phases.__all__
